header: update layout
