+++ src/components/Header/index.js
@@ -1,104 +1,81 @@
 import React from 'react'
 import { themr } from 'react-css-themr'
 import PropTypes from 'prop-types'
-import CloseIcon from 'react-icons/lib/io/android-close'
-import BackIcon from 'react-icons/lib/io/android-arrow-back'
-import {
-  Grid,
-  Row,
-  Col,
-} from 'former-kit'
 
-import Button from '../Button'
+import {
+  length,
+  prop,
+  propEq,
+  slice,
+  findIndex,
+  filter,
+} from 'ramda'
+
+import { ProgressBar } from '..'
 
 const applyThemr = themr('UIHeader')
 
-const buttonColSize = 1
-const imgColSize = 10
+const filterSteps = (pages, active) => {
+  const visibleSteps = filter(prop('visible'), pages)
+  const activeStepIndex = findIndex(propEq('page', active), pages) + 1
+  const activeSteps = slice(1, activeStepIndex, pages)
 
-const buttonPalmColSize = 3
-const imgPalmColSize = 6
+  const activeStep = filter(prop('visible'), activeSteps).length
+  const percentage = (100 / length(visibleSteps)) * (activeStep + 1)
+
+  return {
+    percentage,
+    activeStepIndex: activeStep,
+    filteredSteps: visibleSteps,
+  }
+}
 
 const Header = ({
   logoSrc,
   logoAlt,
-  onPrev,
-  onClose,
-  prevButtonDisabled,
+  steps,
+  activeStep,
   theme,
   base,
-}) => (
-  <header className={theme.header}>
-    <Grid>
-      <Row>
-        <Col
-          tv={buttonColSize}
-          desk={buttonColSize}
-          tablet={buttonColSize}
-          palm={buttonPalmColSize}
-        >
-          <Button
-            base={base}
-            className={theme.back}
-            fill="clean"
-            onClick={onPrev}
-            relevance="low"
-            disabled={prevButtonDisabled}
-            hidden={prevButtonDisabled}
-          >
-            <BackIcon />
-          </Button>
-        </Col>
-        <Col
-          tv={imgColSize}
-          desk={imgColSize}
-          tablet={imgColSize}
-          palm={imgPalmColSize}
-          align={'center'}
-        >
-          {
-            logoSrc && <img
-              className={theme.logo}
-              src={logoSrc}
-              alt={logoAlt}
-            />
-          }
-        </Col>
-        <Col
-          tv={buttonColSize}
-          desk={buttonColSize}
-          tablet={buttonColSize}
-          palm={buttonPalmColSize}
-          align={'end'}
-        >
-          <Button
-            base={base}
-            className={theme.close}
-            fill="clean"
-            onClick={onClose}
-            relevance="high"
-          >
-            <CloseIcon />
-          </Button>
-        </Col>
-      </Row>
-    </Grid>
-  </header>
-)
+}) => {
+  const {
+    percentage,
+    activeStepIndex,
+    filteredSteps,
+  } = filterSteps(steps, activeStep)
+
+  return (
+    <header className={theme.header}>
+      {
+        logoSrc
+          ? <img
+            className={theme.logo}
+            src={logoSrc}
+            alt={logoAlt}
+          />
+          : <span />
+      }
+      <ProgressBar
+        base={base}
+        percentage={percentage}
+        steps={filteredSteps}
+        activeStepIndex={activeStepIndex}
+      />
+    </header>
+  )
+}
+
 
 Header.propTypes = {
   theme: PropTypes.shape({
-    close: PropTypes.string,
     logo: PropTypes.string,
-    back: PropTypes.string,
     header: PropTypes.string,
   }),
   base: PropTypes.string,
   logoSrc: PropTypes.string,
   logoAlt: PropTypes.string,
-  onPrev: PropTypes.func,
-  onClose: PropTypes.func,
-  prevButtonDisabled: PropTypes.bool,
+  steps: PropTypes.arrayOf(PropTypes.object),
+  activeStep: PropTypes.string,
 }
 
 Header.defaultProps = {
@@ -106,9 +83,8 @@ Header.defaultProps = {
   base: 'dark',
   logoSrc: '',
   logoAlt: '',
-  onPrev: null,
-  onClose: null,
-  prevButtonDisabled: false,
+  steps: [],
+  activeStep: '',
 }
 
 export default applyThemr(Header)
