+++ src/themes/default/colors.css
@@ -18,6 +18,7 @@
   --color-grey-83: #2b2b2b;
   --color-grey-84: #282828;
   --color-grey-85: #1f1f1f;
+  --color-grey-86: #242424;
 
   --color-error: #dc6860;
   --color-success: var(--checkout-primary-color);
