+++ src/themes/default/header/index.css
@@ -1,50 +1,16 @@
-@import "../media.css";
-@import "../spacing.css";
 @import "../colors.css";
 
-.close,
-.back {
-  font-size: 30px;
-  padding: 0;
-  margin-top: 22px;
-}
-
-.back {
-  color: var(--color-white);
-}
-
 .logo {
-  max-width: 400px;
-  max-height: 26px;
-  margin-top: 31px;
+  max-width: 110px;
+  max-height: 21px;
+  margin: 29px auto 0;
 }
 
 .header {
-  min-height: 90px;
-  max-height: 90px;
-  margin-bottom: var(--spacing-tiny);
   display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  min-height: 110px;
+  background: var(--color-grey-86);
   width: 100%;
-}
-
-.header > * {
-  align-items: center;
-  flex: 1;
-}
-
-@media (--media-palm) {
-
-  .close,
-  .back {
-    font-size: 35px;
-    margin-top: 20px;
-  }
-
-  .close {
-    margin-right: 22px;
-  }
-
-  .back {
-    margin-left: 22px;
-  }
 }
